Added rcbau::zuul to the repo.
rcbau::zuul opens up the gearman port publically as we will be
building servers, this will be to be secured someone.
But this will need to be discussed, e.g. should we only let
gear listen on the TH_private network (as puppet and the dataset
syncing already do).

If you checkout the zuul node in manifests/site.pp, you'll notice
that I have commented out the stats line, as I assume stats will be
nice, but a graphie server hasn't been created yet.
Further, I have comment out the url_pattern link, is this required?