Change daemonified puppet agent to a crontab entry
There is a bug in puppet agent where it is having trouble handling sshkeys,
which also ends up killing the daemon stopping servers connecting to the
puppetmaster. To get around this, this change creates a crontab entry which
calls the 'run_puppet.sh' script every 10 minutes.

This script runs the normal puppet command, execpt only runs it if the file
'/root/NO_PUPPET' does _NOT_ exist. The rcbau::server now takes a flag 'stop_puppet'
which will create this file on the server if true.

This parameter has been passed up to the other classes so we can turn off puppet on a
node by specifying it in site.pp.